Move sample data to separate file
